// src/app/_dto/TblHPFormat.dto.ts
import { TbldPFormatCo } from "./TbldPFormatCo.dto";

export class TblHPFormat {

    HPFormat_SysID: number;
    HPFormat_Code: string;
    HPFormat_Name: string;
    HPFormat_AcDe_Yn: string;
    HPFormat_AcDe_SysID: number;
    HPFormat_AcDe_Code: string;
    HPFormat_AcDe_Name: string;
    
    TbldPFormatCo: TbldPFormatCo[] = []

}